script to evaluate importance of spikes
I shuffle the spikes corresponding to each neurons in a sliding
window and compute the output of the critic. I then compute the
 absolute difference between the shuffed and non-shuffled samples.
To help visualization I do this for specific samples only presenting
patterns for one randomly selected group of neurons and in a specific
time period of the sample. commented code assess the output of the
critic for different batch sizes.